girar-task-run: avoid one of the races leaving the task POSTPONED forever
The race that is treated by this change:

1. check_depends
2. Another task is done (one of the dependencies of this task), but
has nothing to awake yet, because this task is not yet in POSTPONED
state.
3. This task's state is changed to POSTPONED

...and it is never awaken.

Another race that is not treated:

0. This script locks this task.
1. This task's state is changed to POSTPONED.
2. Another task is done (one of the dependencies of this task), but
cannot awake this task, because this task is still locked by this script.
3. This script finishes and unlocks this task.

The second race could be treated by using lockf -x in gb-y-awake-tasks
rather than lockf -n.